Store current value in MiniWatches
Previously, the current value was queried live from the core and never stored.
Now, it's queried once per search and stored, speeding up further operations with that value.
This allows merging the detailed and normal watch classes into one, as there is no additional
cost in the detailed watch anymore.

The rest of the changeset is related to updating the current value properly, removing IMiniWatchDetailed
references and using the stored current value instead of querying it using GetValue.